feat: Adds basic hello world functionality
Adds the file lib/hello_world_server.rb which allows the
user to visit the specific port in a browser while the ruby program
is running from the command line. Every time the user refreshes the
page while the program is still running, the index will increment
by one and the message "Hello, World! (i)" where i is the current
index will be displayed to the user. Not actually sure this meets
all the requirements of iteration one.